+++ slamrs-robot-rtic/firmware/src/main.rs
@@ -1,6 +1,9 @@
 #![no_main]
 #![no_std]
 
+mod tasks;
+mod util;
+
 // use rp_pico::hal as _;
 use defmt_rtt as _;
 use panic_probe as _;
@@ -15,11 +18,12 @@ use panic_probe as _;
     peripherals = true
 )]
 mod app {
+    use crate::tasks::esp::{init_esp, uart1_esp32};
 
     use defmt::{debug, error, info, warn};
     use embedded_hal::digital::v2::{OutputPin, ToggleableOutputPin};
     use library::event::Event;
-    use library::parse_at::{AtParser, EspMessage, ParsedMessage};
+    use library::parse_at::{AtParser, EspMessage};
 
     use rp_pico::hal::{
         self, clocks,
@@ -36,7 +40,6 @@ mod app {
     };
     use rp_pico::XOSC_CRYSTAL_FREQ;
     use rtic_monotonics::rp2040::*;
-    use rtic_sync::channel::{Sender, TrySendError};
 
     type Uart1Pins = (
         hal::gpio::Pin<Gpio4, hal::gpio::FunctionUart, hal::gpio::PullNone>,
@@ -45,7 +48,7 @@ mod app {
 
     const ESP_CHANNEL_CAPACITY: usize = 32;
     const EVENT_CHANNEL_CAPACITY: usize = 32;
-    type EspChannelReceiver =
+    pub type EspChannelReceiver =
         rtic_sync::channel::Receiver<'static, EspMessage, ESP_CHANNEL_CAPACITY>;
 
     // Shared resources go here
@@ -208,145 +211,34 @@ mod app {
         // continously read the data events and parse them to generate events for the event loop
     }
 
-    /// Task that initializes and handles the ESP wifi connection
-    #[task(
-        priority = 1,
-        local = [
-            esp_mode,
-            esp_reset,
-            uart1_tx,
-            esp_receiver,
-            esp_event_sender,
-        ],
-    )]
-    async fn init_esp(cx: init_esp::Context) {
-        info!("Reseting the ESP");
-        cx.local.esp_mode.set_high().ok();
-        cx.local.esp_reset.set_low().ok();
+    extern "Rust" {
+        // Task that initializes and handles the ESP WIFI connection
+        #[task(
+            priority = 1,
+            local = [
+                esp_mode,
+                esp_reset,
+                uart1_tx,
+                esp_receiver,
+                esp_event_sender,
+            ],
+        )]
+        async fn init_esp(_: init_esp::Context);
 
-        Timer::delay(1.secs()).await;
-        cx.local.esp_reset.set_high().ok();
-        Timer::delay(1.secs()).await;
-
-        // read messages from the device and advance the inner state machine
-
-        wait_for_message(cx.local.esp_receiver, EspMessage::Ready).await;
-
-        // configure some stuff
-        cx.local.uart1_tx.write_full_blocking(b"AT+SYSMSG=0\r\n");
-
-        enum State {
-            Ready,
-            WifiConnectedAndIp,
-            Listening,
-            ClientConnected,
-        }
-
-        let mut state = State::Ready;
-
-        info!("Done, starting message loop");
-        loop {
-            while let Ok(m) = cx.local.esp_receiver.recv().await {
-                info!("Got message: {}", m);
-                match m {
-                    EspMessage::GotIP => {
-                        state = State::WifiConnectedAndIp;
-                        // start the server
-
-                        info!("Enabling Multiple Connections");
-                        cx.local.uart1_tx.write_full_blocking(b"AT+CIPMUX=1\r\n");
-                        wait_for_message(cx.local.esp_receiver, EspMessage::Ok).await;
-                        Timer::delay(1.secs()).await;
-
-                        cx.local
-                            .uart1_tx
-                            .write_full_blocking(b"AT+CIPSERVERMAXCONN=1\r\n");
-                        wait_for_message(cx.local.esp_receiver, EspMessage::Ok).await;
-
-                        info!("Starting server");
-                        cx.local
-                            .uart1_tx
-                            .write_full_blocking(b"AT+CIPSERVER=1,80\r\n");
-                        wait_for_message(cx.local.esp_receiver, EspMessage::Ok).await;
-
-                        state = State::Listening;
-                        info!("Listening");
-                    }
-                    EspMessage::ClientConnect => {
-                        state = State::ClientConnected;
-                        channel_send(cx.local.esp_event_sender, Event::Connected, "ESP");
-                    }
-                    EspMessage::ClientDisconnect => {
-                        state = State::Listening;
-                        channel_send(cx.local.esp_event_sender, Event::Disconnected, "ESP");
-                    }
-                    _ => {}
-                }
-            }
-        }
-    }
-
-    async fn wait_for_message(receiver: &mut EspChannelReceiver, value: EspMessage) {
-        while let Ok(m) = receiver.recv().await {
-            if m == value {
-                return;
-            } else {
-                warn!("got message {} while waiting for {}", m, value);
-            }
-        }
-    }
-
-    /// Helper function for trying to send something to a Sender MPSC channel, or print a warning
-    /// message if an error occurred
-    fn channel_send<T: defmt::Format, const N: usize>(
-        sender: &mut Sender<'static, T, N>,
-        value: T,
-        context: &str,
-    ) {
-        match sender.try_send(value) {
-            Err(TrySendError::Full(m)) => {
-                warn!("ESP channel full, failed to send: {} ({})", m, context)
-            }
-            Err(TrySendError::NoReceiver(m)) => {
-                warn!(
-                    "ESP channel has no receiver, failed to send: {} ({})",
-                    m, context
-                )
-            }
-            _ => {}
-        }
-    }
-
-    /// Hardware task that reads bytes from the UART an publishes messages!
-    #[task(
-        binds = UART1_IRQ,
-        local = [
-            uart1_rx,
-            esp_sender,
-            parser: AtParser<256> = AtParser::new(),
-        ],
-    )]
-    fn uart1_esp32(cx: uart1_esp32::Context) {
-        let sender = cx.local.esp_sender;
-        let rx = cx.local.uart1_rx;
-        cx.local.parser.consume(rx, move |message| match message {
-            ParsedMessage::Simple(m) => channel_send(sender, m, "uart1_esp32"),
-            ParsedMessage::ReceivedData(data) => {
-                info!("got data: {}", data);
-            }
-        });
+        // Hardware task that reads bytes from the UART an publishes messages!
+        #[task(
+            binds = UART1_IRQ,
+            local = [
+                uart1_rx,
+                esp_sender,
+                parser: AtParser<256> = AtParser::new(),
+            ],
+        )]
+        fn uart1_esp32(cx: uart1_esp32::Context);
     }
 
     #[task(local = [led])]
     async fn heartbeat(ctx: heartbeat::Context) {
-        // Loop forever.
-        //
-        // It is important to remember that tasks that loop
-        // forever should have an `await` somewhere in that loop.
-        //
-        // Without the await, the task will never yield back to
-        // the async executor, which means that no other lower or
-        // equal  priority task will be able to run.
         loop {
             // Flicker the built-in LED
             _ = ctx.local.led.toggle();
